Single-container API+Frontend+Worker
