+++ api/obs/api/routes/info.py
@@ -2,7 +2,7 @@ import logging
 
 # from sqlalchemy import select
 
-from obs.api.app import app
+from obs.api.app import api
 
 from sanic.response import json
 
@@ -11,7 +11,7 @@ log = logging.getLogger(__name__)
 from obs.api import __version__ as version
 
 
-@app.route("/info")
+@api.route("/info")
 async def info(req):
     return json(
         {
